requirements.txt
# Django and REST Framework
Django>=4.2,<5.0
djangorestframework>=3.14,<4.0
django-cors-headers>=4.3,<5.0

# Database
psycopg2-binary>=2.9,<3.0

# Celery and Redis
celery>=5.3,<6.0
redis>=5.0,<6.0
django-redis>=5.4,<6.0

# Web Scraping
requests>=2.31,<3.0
beautifulsoup4>=4.12,<5.0
lxml>=5.0,<6.0
fake-useragent>=1.4,<2.0

# Environment and Configuration
python-decouple>=3.8,<4.0

# API Documentation
drf-spectacular>=0.27,<1.0

# Rate Limiting
django-ratelimit>=4.1,<5.0

# Testing
pytest>=7.4,<8.0
pytest-django>=4.7,<5.0

# Production
gunicorn>=21.0,<22.0
